Moving RetryRelayer to FullPreset
