+++ crates/relayer-framework/src/base/one_for_all/impls/relay/mod.rs
@@ -2,5 +2,4 @@ pub mod errors;
 pub mod message_builders;
 pub mod message_sender;
 pub mod packet_relayers;
-pub mod retry;
 pub mod types;
