+++ crates/relayer-framework/src/base/one_for_all/instances/packet_relayers.rs
@@ -4,7 +4,8 @@ use crate::base::one_for_all::traits::relay::OfaBaseRelay;
 use crate::base::one_for_all::traits::relay::OfaRelayPreset;
 use crate::base::relay::impls::packet_relayers::ack::base_ack_packet::BaseAckPacketRelayer;
 use crate::base::relay::impls::packet_relayers::general::full_relay::FullRelayer;
-use crate::base::relay::impls::packet_relayers::general::retry::RetryRelayer;
+use crate::full::one_for_all::traits::relay::OfaFullRelay;
+use crate::full::relay::impls::packet_relayers::retry::RetryRelayer;
 use crate::base::relay::impls::packet_relayers::receive::base_receive_packet::BaseReceivePacketRelayer;
 use crate::base::relay::impls::packet_relayers::receive::skip_received_packet::SkipReceivedPacketRelayer;
 use crate::base::relay::traits::packet_relayer::PacketRelayer;
@@ -14,7 +15,7 @@ use crate::common::one_for_all::types::relay::OfaRelayWrapper;
 
 pub fn full_packet_relayer<Relay>() -> PhantomData<impl PacketRelayer<OfaRelayWrapper<Relay>>>
 where
-    Relay: OfaBaseRelay,
+    Relay: OfaFullRelay,
     Relay::Preset: OfaRelayPreset<Relay>,
 {
     PhantomData::<RetryRelayer<FullRelayer>>
