+++ crates/relayer-framework/src/base/relay/impls/packet_relayers/general/mod.rs
@@ -1,2 +1,1 @@
 pub mod full_relay;
-pub mod retry;
