+++ crates/relayer-framework/src/common/one_for_all/presets/mod.rs
@@ -1,3 +1,5 @@
+/// Captures a minimal set of relayer features.
 pub struct MinimalPreset;
 
+/// Captures the full set of relayer features.
 pub struct FullPreset;
