+++ crates/relayer-framework/src/full/mod.rs
@@ -2,4 +2,5 @@ pub mod all_for_one;
 pub mod batch;
 pub mod filter;
 pub mod one_for_all;
+pub mod relay;
 pub mod telemetry;
