+++ crates/relayer-framework/src/full/one_for_all/impls/mod.rs
@@ -2,4 +2,5 @@ pub mod batch;
 pub mod chain;
 pub mod filter;
 pub mod preset;
+pub mod relay;
 pub mod telemetry;
