+++ crates/relayer-framework/src/full/one_for_all/impls/preset.rs
@@ -4,7 +4,7 @@ use crate::base::one_for_all::traits::chain::OfaIbcChain;
 use crate::base::one_for_all::traits::chain::{OfaChainPreset, OfaIbcChainPreset};
 use crate::base::one_for_all::traits::relay::OfaRelayPreset;
 use crate::base::relay::impls::packet_relayers::general::full_relay::FullRelayer;
-use crate::base::relay::impls::packet_relayers::general::retry::RetryRelayer;
+use crate::full::relay::impls::packet_relayers::retry::RetryRelayer;
 use crate::common::one_for_all::presets::FullPreset;
 use crate::full::batch::message_sender::SendMessagetoBatchWorker;
 use crate::full::filter::impls::filter_relayer::FilterRelayer;
