+++ crates/relayer-framework/src/full/one_for_all/impls/relay/retry.rs
@@ -1,8 +1,8 @@
-use crate::base::one_for_all::traits::relay::OfaBaseRelay;
-use crate::base::relay::impls::packet_relayers::general::retry::SupportsPacketRetry;
 use crate::common::one_for_all::types::relay::OfaRelayWrapper;
+use crate::full::one_for_all::traits::relay::OfaFullRelay;
+use crate::full::relay::impls::packet_relayers::retry::SupportsPacketRetry;
 
-impl<Relay: OfaBaseRelay> SupportsPacketRetry for OfaRelayWrapper<Relay> {
+impl<Relay: OfaFullRelay> SupportsPacketRetry for OfaRelayWrapper<Relay> {
     const MAX_RETRY: usize = 3;
 
     fn is_retryable_error(e: &Self::Error) -> bool {
