+++ crates/relayer-framework/src/full/one_for_all/traits/relay.rs
@@ -5,4 +5,8 @@ use async_trait::async_trait;
 #[async_trait]
 pub trait OfaFullRelay: OfaBaseRelay {
     async fn should_relay_packet(&self, packet: &Self::Packet) -> Result<bool, Self::Error>;
+
+    fn is_retryable_error(e: &Self::Error) -> bool;
+
+    fn max_retry_exceeded_error(e: Self::Error) -> Self::Error;
 }
